Colocate funding finder and programmes, add tests
